fix(EventStore): Fixed the ESEventStore by catching WrongExceptionVersionException and transforming them into OptimisticConcurrencyExceptions fix(Core): Move ConcurrencyExceptions from Caching.Abstractions into Core project
